+++ eslint.config.js
@@ -11,6 +11,8 @@ module.exports = [
       "test/cli/fixtures/bad.js", // Intentionally-invalid
       "test/cli/fixtures/imports_peggy.js", // Generated
       "test/cli/fixtures/lib.js", // Generated
+      "test/cli/fixtures/useFrags/fs.js", // Generated
+      "test/cli/fixtures/useFrags/identifier.js", // Generated
       "test/vendor/**",
     ],
   },
@@ -57,4 +59,11 @@ module.exports = [
       "compat/compat": "error",
     },
   },
+  // {
+  //   files: [
+  //   ],
+  //   languageOptions: {
+  //     sourceType: "module",
+  //   },
+  // },
 ];